Moved armorStrengthBreakProbabilityLookup from EnemyInformation to Enemy
It makes more sense there, in the new architecture.

Also, cleaned up TODOs.
