+++ src/main/java/drgtools/dpscalc/modelPieces/statusEffects/MultipleSTEs.java
@@ -13,7 +13,6 @@ import java.util.EnumMap;
 // Needs to add up the DPS, split by DMG_Element, aggregate Heat/sec, multiply the slows, etc etc
 // Also needs to transparently work if there's only one STE, because that way it can be used as an entry point for Breakpoints()
 public class MultipleSTEs {
-    // TODO: I think this is only used in Breakpoints?
     private ArrayList<PushSTEComponent> statusEffects;
     private EnumMap<DamageElement, Double> totalDamageByElement;
     private double complicatedSlowUtilitySum;
